+++ src/pipls/__init__.py
@@ -1,7 +1,8 @@
 """Pi-PLS multivariate regression."""
 
+from .exceptions import StatisticalSupportWarning
 from .regression import PiPLSRegression
 
-__all__ = ["PiPLSRegression", "__version__"]
+__all__ = ["PiPLSRegression", "StatisticalSupportWarning", "__version__"]
 
 __version__ = "0.0.0"